saving access after paystack verification

+++ ImoLandRegistry/src/main/java/com/imo/landregistry/dao/PaymentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.imo.landregistry.entity.PaymentEntity;
 
-public interface PaymentRepository extends JpaRepository<PaymentEntity, Long> {
+public interface PaymentRepository extends JpaRepository<PaymentEntity, Integer> {
 
 }
